Remove select_age module
select_age module contains one method that can be included
in resources.

In my opinion it should be contained in resources since
there is no real use for it as a seperate object. It only
selects resources (as resource module) but by age parameters.

In addition, it will allow us to move all "old" and "unused" logic
from janitor main loop to resources module where it should be.
